Move billing API setup into go.billing.api.
